+++ js/controllers/gegectrl/manager-operate.js
@@ -1,38 +1,60 @@
 'use strict';
 
 app
-	// Flot Chart controller
-	.controller('ManagerOperateCtrl', ['$scope', 'commonService', 'modelService', function($scope, commonService, modelService) {
+	.controller('ManagerOperateCtrl', ['$scope', 'commonService', 'modelService', '$newLocalStorage', '$state', function($scope, commonService, modelService, $newLocalStorage, $state) {
 
 		$scope.currentPageNo = 1;
 		$scope.pageSize = 16;
 		$scope.managerList = [];
 		$scope.managerInfo = {};
 		$scope.pageList = [];
+		$scope.showContent = false; //是否显示内容
+		$scope.disableHospital = true; //禁用医院选择
+		$scope.lockInput = false;
 		$scope.initManagerInfo = {
 
 		};
+		$scope.gegeUser = JSON.parse($newLocalStorage.get('gege_manager'));
+		if($scope.gegeUser) {
+			//判断是否有权限
+			if($scope.$parent.gegePermisson.managerPermisson) {
+				$scope.showContent = true;
+			} else {
+				alert('您还没有此权限，联系管理员开通吧');
+				return;
+			}
+		} else {
+			$state.go('access.signin');
+		}
+
+		console.log($scope.gegeUser);
+		//获取管理员权限列表
+		$scope.getPermissionList = function() {
+			$scope.permissionList = angular.copy($scope.gegeUser.Admpermissionlist);
+			console.log($scope.permissionList);
+			$scope.permissionList = _.map($scope.permissionList, function(item) {
+				item.checked = false;
+				return item;
+			});
+			console.log($scope.permissionList);
+		}
+		$scope.getPermissionList();
+		//判断是平台还是院内管理员
+		if($scope.gegeUser.Role == 1) {
+			$scope.disableHospital = false;
+		}
+
 		//根据分页获取管理员列表
 		$scope.getManagerInfoList = function(page) {
 			modelService.getAdmin({
 				pageNumber: page,
-				pageSize: $scope.pageSize
+				pageSize: $scope.pageSize,
+				adminId: $scope.gegeUser.AdmId
 			}).then(function(res) {
 				if(res.code == 0) {
 					//处理返回数据
 					$scope.managerList = _.map(commonService.translateServerData(res.body), function(item) {
-						item.OperatorTime = commonService.str2date(item.OperatorTime, 'yyyy-MM-dd');
-						for(var i = 0; i < item.Admpermissionlist.length; i++) {
-							if(item.Admpermissionlist[i].PermissionId == '000000001') {
-								item.isblsj = true;
-							}
-							if(item.Admpermissionlist[i].PermissionId == '000000002') {
-								item.ispb = true;
-							}
-							if(item.Admpermissionlist[i].PermissionId == '000000003') {
-								item.isxf = true;
-							}
-						}
+						item.OpTime = commonService.str2date(item.OperatorTime, 'yyyy-MM-dd');
 						return item;
 					});
 					$scope.pageList = [];
@@ -45,7 +67,6 @@ app
 						$scope.pageList.push(i);
 					}
 				} else {
-					alert('数据为空');
 					$scope.pageList = [];
 				}
 			}, function(err) {});
@@ -100,12 +121,16 @@ app
 		//根据医院改变科室
 		$scope.changeHospital = function(hospitalId) {
 			//调用获取科室服务
+			console.log(hospitalId);
 			modelService.getDepByHospId({
 				HospitalId: hospitalId
 			}).then(function(res) {
 				console.log(res);
 				if(res.code == 0) {
-					$scope.depList = res.body;
+					$scope.depList = _.map(res.body, function(item) {
+						item.DepartmentChecked = false;
+						return item;
+					});
 					if($scope.operateState == 'edit') {
 						for(var i = 0; i < $scope.depList.length; i++) {
 							for(var j = 0; j < $scope.managerInfo.Admdepartmentlist.length; j++) {
@@ -114,15 +139,21 @@ app
 								}
 							}
 						}
-						
+
+					}
+					if($scope.operateState == 'add'&&$scope.gegeUser.Role == 1) {
+						$scope.managerInfo.selectAll = true;
+						$scope.lockDep = true;
+						$scope.selectAll();
 					}
 				}
 
 			}, function(err) {});
 		}
 
-		//全选
+		//全选科室
 		$scope.selectAll = function() {
+			console.log('all');
 			for(var i = 0; i < $scope.depList.length; i++) {
 				if($scope.managerInfo.selectAll === true) {
 					$scope.depList[i].DepartmentChecked = true;
@@ -135,31 +166,39 @@ app
 		//添加管理员
 		$scope.addManager = function() {
 			$scope.operateState = 'add';
-			$scope.depList=[];
+			$scope.lockInput = false; //打开输入框
+			if($scope.gegeUser.Role == 0 || $scope.gegeUser.Role == 2) {
+				$scope.disableHospital = true; //禁用医院
+			} else if($scope.gegeUser.Role == 1) {
+				$scope.disableHospital = false; //可选医院
+			}
+			$scope.depList = [];
 			$('tbody tr').removeClass('tr-success');
 			$scope.selectData = false;
 			$('#modal_showAudit').modal('show');
 			$scope.managerInfo = angular.copy($scope.initManagerInfo);
 			$scope.managerInfo.Admpermissionlist = [];
 			$scope.managerInfo.Admdepartmentlist = [];
-
+			$scope.getPermissionList();
 		}
 
 		//删除管理员
 		$scope.deleteManager = function() {
 			if($scope.selectData) {
-				modelService.deleteAdmin({
-					AdmId: $scope.managerInfo.AdmId
-				}).then(function(res) {
-					if(res.code == 0) {
-						alert('删除成功');
-						$scope.getManagerInfoList($scope.currentPageNo);
-					} else {
-						alert('删除失败');
-					}
-				}, function(err) {
-
-				});
+				if(confirm("确定要删除该账号么？")) {
+					modelService.deleteAdmin({
+						AdmId: $scope.managerInfo.AdmId
+					}).then(function(res) {
+						if(res.code == 0) {
+							alert('删除成功');
+							$scope.getManagerInfoList($scope.currentPageNo);
+						} else {
+							alert('删除失败');
+						}
+					}, function(err) {
+						alert('网络错误！');
+					});
+				}
 			} else {
 				alert('请选择要删除的管理员');
 			}
@@ -169,58 +208,70 @@ app
 		//显示编辑框
 		$scope.editManager = function() {
 			$scope.operateState = 'edit';
+			$scope.lockInput = true; //禁用输入框
+			if($scope.gegeUser.Role == 0 || $scope.gegeUser.Role == 2) {
+				$scope.disableHospital = true; //禁用医院
+			} else if($scope.gegeUser.Role == 1) {
+				$scope.disableHospital = false; //可选医院
+			}
 			if($scope.selectData) {
 				console.log($scope.managerInfo);
 				$('#modal_showAudit').modal('show');
 				$scope.changeHospital($scope.managerInfo.HospitalId);
-				console.log($scope.depList);
+				for(var i = 0; i < $scope.permissionList.length; i++) {
+					for(var j = 0; j < $scope.managerInfo.Admpermissionlist.length; j++) {
+						if($scope.permissionList[i].PermissionId == $scope.managerInfo.Admpermissionlist[j].PermissionId) {
+							$scope.permissionList[i].checked = true;
+						}
+					}
+				}
 			} else {
 				alert('请选择要编辑的管理员');
 			}
 
 			console.log($scope.managerInfo);
 		}
+
+		//处理权限数据
+		$scope.dealPmnAndDep = function() {
+			$scope.managerInfo.OperatorId = $scope.gegeUser.AdmId;
+			$scope.managerInfo.Admpermissionlist = [];
+			$scope.managerInfo.Admdepartmentlist = [];
+			for(var j = 0; j < $scope.permissionList.length; j++) {
+				if($scope.permissionList[j].checked) {
+					var obj = {};
+					obj.PermissionId = $scope.permissionList[j].PermissionId;
+					obj.PermissionName = $scope.permissionList[j].PermissionName;
+					$scope.managerInfo.Admpermissionlist.push(obj);
+				}
+			}
+			//处理选择的科室
+			for(var i = 0; i < $scope.depList.length; i++) {
+				if($scope.depList[i].DepartmentChecked) {
+					var obj = {
+						DepartmentId: $scope.depList[i].DepartmentId
+					};
+					$scope.managerInfo.Admdepartmentlist.push(obj);
+				}
+			}
+		}
 		//更新管理员
 		$scope.subManagerInfo = function(item) {
 			if($scope.operateState == 'add') {
-				//处理不良事件、排版和学分的选择
-				if($scope.managerInfo.isblsj) {
-					var obj1 = {
-						AdmId: $scope.managerInfo.UserId,
-						PermissionId: "000000001"
-					}
-					$scope.managerInfo.Admpermissionlist.push(obj1);
+				if($scope.gegeUser.Role == 0) {
+					$scope.managerInfo.Role = 1;
+				} else if($scope.gegeUser.Role == 1 || $scope.gegeUser.Role == 2) {
+					$scope.managerInfo.Role = 2;
 				}
-				if($scope.managerInfo.ispb) {
-					var obj2 = {
-						AdmId: $scope.managerInfo.UserId,
-						PermissionId: "000000002"
-					}
-					$scope.managerInfo.Admpermissionlist.push(obj2);
-				}
-				if($scope.managerInfo.isxf) {
-					var obj3 = {
-						AdmId: $scope.managerInfo.UserId,
-						PermissionId: "000000003"
-					}
-					$scope.managerInfo.Admpermissionlist.push(obj3);
-				}
-				//处理选择的科室
-				for(var i = 0; i < $scope.depList.length; i++) {
-					if($scope.depList[i].DepartmentChecked) {
-						var obj = {
-							AdmId: $scope.managerInfo.UserId,
-							DepartmentId: $scope.depList[i].DepartmentId
-						};
-						$scope.managerInfo.Admdepartmentlist.push(obj);
-					}
-				}
+				$scope.dealPmnAndDep();
 				var data = JSON.stringify({
 					model: $scope.managerInfo
 				});
+				console.log(data);
 				modelService.addAdmin(data).then(function(res) {
 					if(res.code == 0) {
 						alert('添加成功');
+						$scope.getManagerInfoList($scope.currentPageNo);
 					} else {
 						alert('添加失败');
 					}
@@ -230,42 +281,14 @@ app
 					alert('网络出错，请刷新重试！');
 				});
 			} else if($scope.operateState == 'edit') {
-				if($scope.managerInfo.isblsj) {
-					var obj1 = {
-						AdmId: $scope.managerInfo.UserId,
-						PermissionId: "000000001"
-					}
-					$scope.managerInfo.Admpermissionlist.push(obj1);
-				}
-				if($scope.managerInfo.ispb) {
-					var obj2 = {
-						AdmId: $scope.managerInfo.UserId,
-						PermissionId: "000000002"
-					}
-					$scope.managerInfo.Admpermissionlist.push(obj2);
-				}
-				if($scope.managerInfo.isxf) {
-					var obj3 = {
-						AdmId: $scope.managerInfo.UserId,
-						PermissionId: "000000003"
-					}
-					$scope.managerInfo.Admpermissionlist.push(obj3);
-				}
-				for(var i = 0; i < $scope.depList.length; i++) {
-					if($scope.depList[i].DepartmentChecked) {
-						var obj = {
-							AdmId: $scope.managerInfo.UserId,
-							DepartmentId: $scope.depList[i].DepartmentId
-						};
-						$scope.managerInfo.Admdepartmentlist.push(obj);
-					}
-				}
+				$scope.dealPmnAndDep();
 				var data = JSON.stringify({
 					model: $scope.managerInfo
 				});
 				modelService.updateAdmin(data).then(function(res) {
 					if(res.code == 0) {
 						alert('更新成功');
+						$scope.getManagerInfoList($scope.currentPageNo);
 					} else {
 						alert('更新失败');
 					}
@@ -276,12 +299,38 @@ app
 				$('#modal_showAudit').modal('hide');
 			}
 
-			//			$scope.managerInfo = {};
 		}
+		//重置管理员密码
+		$scope.restPwd = function() {
+			if($scope.selectData) {
+				if(confirm("确定要重置密码么？")) {
+					console.log(JSON.stringify({
+						model: $scope.managerInfo
+					}));
+
+					modelService.resetAdminPassword({
+						model: $scope.managerInfo
+					}).then(function(res) {
+						console.log(res);
+						if(res.code == 0) {
+							alert('密码已被重置为：' + res.body.Password + ',请尽快修改！');
+							$scope.getManagerInfoList($scope.currentPageNo);
+						} else {
+							alert('密码重置失败！');
+						}
+					}, function(err) {
+						alert('网络错误！刷新重试')
+					});
+				}
+
+			} else {
+				alert('请选择要重置密码的管理员');
+			}
+		}
+
 		//关闭编辑框
 		$scope.closeModal = function() {
 			$('#modal_showAudit').modal('hide');
-			$scope.managerInfo = angular.copy($scope.initManagerInfo);
 		}
 
 		//刷新页面
